HeaderField position from u16 to u64, some refactoring

+++ src/headers/exth.rs
@@ -19,7 +19,7 @@ pub(crate) enum ExtHeaderData {
     RecordCount,
 }
 impl HeaderField for ExtHeaderData {
-    fn position(self) -> u16 {
+    fn position(self) -> u64 {
         match self {
             ExtHeaderData::Identifier => 328,
             ExtHeaderData::HeaderLength => 332,
